+++ tests/utils.py
@@ -1,4 +1,5 @@
 import re
+import os
 import sys
 import time
 import warnings
@@ -78,14 +79,18 @@ def id_sleep(x, delay=0):
 
 
 def check_subprocess_call(cmd, timeout=1, stdout_regex=None,
-                          stderr_regex=None):
+                          stderr_regex=None, env=None):
     """Runs a command in a subprocess with timeout in seconds.
 
     Also checks returncode is zero, stdout if stdout_regex is set, and
     stderr if stderr_regex is set.
     """
+    if env is not None:
+        env_ = os.environ.copy()
+        env_.update(env)
+        env = env_
     proc = subprocess.Popen(cmd, stdout=subprocess.PIPE,
-                            stderr=subprocess.PIPE)
+                            stderr=subprocess.PIPE, env=env)
 
     def kill_process():
         warnings.warn("Timeout running {}".format(cmd))